# src/main/resources/AwsServicesOpenApi.yaml
openapi: 3.0.1
info:
  title: AwsServicesTest
  description: This APi test or implements diffrent AWS services
  termsOfService: NA
  contact:
    email: er.vikaspaul@gmail.com
  license:
    name: AwsServices
  version: 1.0.0
...
servers:
  - url: 'https://localhost:7070/awsservices/v1'
...
paths:
  /customer:
    put:
      tags:
        - customer
      summary: Update an existing customer
      description: Update an existing known customer
      operationId: updateCustomer
      requestBody:
        description: customer update request
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerUpdateRequest'
        required: true
      responses:
        '202':
          description: Customer Sucessfully Updated
        '400':
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '405':
          description: Validation exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        default:
          description: successful operation
          content: {}
    post:
      tags:
        - customer
      summary: Add new Customer
      description: Add new customer which has admin role
      operationId: addCustomer
      requestBody:
        description: enter the appropriate customer creation Request
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerCreationRequest'
        required: true
      responses:
        '202':
          description: Customer Sucessfully Added
        '400':
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '405':
          description: Validation exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        default:
          description: successful operation
          content: {}
  '/customer/{customerId}':
    get:
      tags:
        - customer
      summary: Find customer by ID
      description: Returns a single customer
      operationId: getCustomerById
      parameters:
        - name: customerId
          in: path
          description: ID of customer to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Customer Details Fetched Successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerDetailsResponse'
        '400':
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '405':
          description: Validation exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        default:
          description: successful operation
          content: {}
    delete:
      tags:
        - customer
      summary: Deletes a Customer
      description: Deletes a Customer and underlying details
      operationId: deleteCustomerById
      parameters:
        - name: customerId
          in: path
          description: ID of customer to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '202':
          description: Customer Deleted Sucessfully
        '400':
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '405':
          description: Validation exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        default:
          description: successful operation
          content: {}
...
components:
  schemas:
    CustomerCreationRequest:
      type: object
      required: 
        - companyName
        - contactName
        - contactEmail
        - contactNumber
        - type
      properties:
        companyName:
          description: Name of the Company of the Registering Customer
          type: string
        contactName:
          description: Contact person of thest company
          items:
            $ref: '#/components/schemas/CompanyContactEmail'
        contactEmail:
          description: Company Contact Email
          items:
            $ref: '#/components/schemas/CompanyContactEmail'
        contactNumber:
          description: Contact number of the customer
          items:
            $ref: '#/components/schemas/CompanyNumber'
        type:
          description: Type of Customer
          items:
            $ref: '#/components/schemas/CustomerType'
    CustomerUpdateRequest:
      type: object
      required: 
        - id
        - companyName
        - contactName
        - contactEmail
        - contactNumber
        - type
      properties:
        id:
          description: id of the customer
          type: integer
          format: int32
        companyName:
          description: Name of the Company of the Registering Customer
          items:
            $ref: '#/components/schemas/CompanyName'
        contactName:
          description: Contact person of thest company
          items:
            $ref: '#/components/schemas/CompanyContactName'
        contactEmail:
          description: Company Contact Email
          items:
            $ref: '#/components/schemas/CompanyContactEmail'
        contactNumber:
          description: Contact number of the customer
          items:
            $ref: '#/components/schemas/CompanyNumber'
        type:
          description: Type of Customer
          items:
            $ref: '#/components/schemas/CustomerType'
        hsmUserId:
          description: Hsm Customer ID
          items:
            $ref: '#/components/schemas/HsmUserID'
        publicKeyConfirmed:
          description: Public Key Confirmed
          items:
            $ref: '#/components/schemas/PublicKeyConfirmed'
        customerUUID:
          description: Customer UUID
          items:
            $ref: '#/components/schemas/CustomerUUID'
    CustomerDetailsResponse:
      type: object
      properties:
        id:
          description: id of the customer
          type: integer
          format: int64
        companyName:
          description: Name of the Company of the Registering Customer
          items:
            $ref: '#/components/schemas/CompanyName'
        contactName:
          description: Contact person of thest company
          items:
            $ref: '#/components/schemas/CompanyContactName'
        contactEmail:
          description: Company Contact Email
          items:
            $ref: '#/components/schemas/CompanyContactEmail'
        contactNumber:
          description: Contact number of the customer
          items:
            $ref: '#/components/schemas/CompanyNumber'
        type:
          description: Type of Customer
          items:
            $ref: '#/components/schemas/CustomerType'
        hsmUserId:
          description: Hsm Customer ID
          items:
            $ref: '#/components/schemas/HsmUserID'
        publicKeyConfirmed:
          description: Public Key Confirmed
          items:
            $ref: '#/components/schemas/PublicKeyConfirmed'
        customerUUID:
          description: Customer UUID
          items:
            $ref: '#/components/schemas/CustomerUUID'
...
    ApiErrorResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        type:
          type: string
        message:
          type: string
    CompanyName:
      type: string
      maxLength: 50
      minLength: 1
    CompanyContactName:
      type: string
      maxLength: 50
      minLength: 1
    CompanyContactEmail:
      type: string
    CompanyNumber:
      type: integer
      format: int32
      minimum: 1111111111
    CustomerType:
      type: string
      enum:
        - Small
        - Medium
        - Large
        - Dummy
    HsmUserID:
      type: string
    PublicKeyConfirmed:
      type: boolean
      default: false
    CustomerUUID:
      type: string
...